vic: fix spriteLine overflow for sprites in the border
The limit of sprite's X coordinate to SPRITE_MAX_X means part of sprite
might be outside SPRITE_MAX_X, which is how spriteLine is sized. The
loops processing spriteData must not bump and use the slp pointer
outside spriteLine, but that can currently happen with the fixed 24
limit for "cnt". Make sure it's limited based on (SPRITE_MAX_X - x)
divided accordingly to how many slp increments happen in one iteration.

This fixes a bug in River Raid where horizontally flying white planes
would disappear as if shooted when they are supposed to cross and wrap
over the border, during which the X position is temporarily just below
SPRITE_MAX_X.  That was caused by a spurious collision detected due to
reading improper data outside of spriteLine, thus lineMemChr. Writing
the plane's color data there was also likely corrupting lineMemChr too.

Also limit the iterations from 24 to 12 for multicolor sprites as it
processes spriteData two bits per iteration. This was harmless though as
the loop would be terminated by the (spriteData > 0) condition becoming
false.

Also add a TODO for the SPRITE_MAX_X define as AFAIU it's not correct to
limit sprite processing (mostly the collision detection) to this
particular x coordinate. But I need to read the docs closer.